refactor(kitchenpos-menu): menuProduct와 menu를 애그리거트로 묶는다
