+++ src/main/java/kitchenpos/menus/ui/dto/MenuCreateRequest.java
@@ -6,7 +6,7 @@ import kitchenpos.menus.domain.tobe.menu.DisplayedMenu;
 import kitchenpos.menus.domain.tobe.menu.Menu;
 import kitchenpos.menus.domain.tobe.menu.MenuName;
 import kitchenpos.menus.domain.tobe.menu.MenuPrice;
-import kitchenpos.menus.domain.tobe.menuproduct.MenuProducts;
+import kitchenpos.menus.domain.tobe.menu.MenuProducts;
 import kitchenpos.products.domain.ProfanityValidator;
 
 public class MenuCreateRequest {
