+++ src/test/java/kitchenpos/fake/InMemoryMenuRepository.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.menus.domain.tobe.menu.Menu;
-import kitchenpos.menus.domain.tobe.menuproduct.MenuProduct;
+import kitchenpos.menus.domain.tobe.menu.MenuProduct;
 
 public class InMemoryMenuRepository implements MenuRepository {
