+++ src/test/java/kitchenpos/fixture/MenuProductFixture.java
@@ -1,8 +1,8 @@
 package kitchenpos.fixture;
 
 import java.util.List;
-import kitchenpos.menus.domain.tobe.menuproduct.MenuProduct;
-import kitchenpos.menus.domain.tobe.menuproduct.MenuQuantity;
+import kitchenpos.menus.domain.tobe.menu.MenuProduct;
+import kitchenpos.menus.domain.tobe.menu.MenuQuantity;
 import kitchenpos.menus.ui.dto.MenuProductCreateRequest;
 import kitchenpos.menus.ui.dto.MenuProductCreateRequests;
 import kitchenpos.products.domain.tobe.Product;
