+++ src/test/java/kitchenpos/menus/domain/tobe/menu/MenuProductsTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain.tobe.menuproduct;
+package kitchenpos.menus.domain.tobe.menu;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -6,6 +6,8 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.fixture.MenuProductFixture;
 import kitchenpos.menus.domain.tobe.menu.MenuPrice;
+import kitchenpos.menus.domain.tobe.menu.MenuProduct;
+import kitchenpos.menus.domain.tobe.menu.MenuProducts;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
